requirements.txt
anyio==3.6.2
async-exit-stack==1.0.1
async-generator==1.10
asyncpg==0.27.0
click==8.1.3
databases==0.7.0
fastapi~=0.95.1
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
psycopg2-binary==2.9.6
pydantic==1.10.7
python-dotenv==1.0.0
PyYAML==6.0
sniffio==1.3.0
SQLAlchemy~=2.0.12
starlette==0.27.0
typing_extensions==4.5.0
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3

jose~=1.0.0
passlib~=1.7.4
sendgrid~=6.10.0